truncate 적용, videoitem 수정
